DjangoAPI/requirement.txt
absl-py==1.3.0
astor==0.8.1
Django>=4.2
django-crispy-forms>=2.3
djangorestframework==3.13.1
gast==0.4.0
google-pasta==0.2.0
grpcio>=1.48.2
gunicorn==20.1.0
h5py==3.6.0
joblib>=1.1.1
Keras==2.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.3
numpy==1.24.2
pandas==1.5.3
protobuf==4.23.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
tensorboard==2.14.0
tensorflow==2.14.0
tensorflow-estimator==2.14.0
termcolor==2.3.0
Werkzeug==2.2.3
whitenoise==6.1.0
wrapt==1.14.1
xlrd==2.0.1
